Avoid wait forever from t/pod/pod2usage2.t
Fixes #10

We need to be sure to use the last version of
several Pod modules in order to avoid issues.

Also note that perldoc is by default not available
on ubuntu containers and points to a fake binary.

This is adding an extra workflow to check the two
scenario.

This is also fixing a warning from the windows workflow.