remove vendored Pod::Parser and port test to use Pod::Text
One of the tests was testing the result of running the pod2usage.PL file
though Pod::PlainText. Because Pod::PlainText was part of Pod::Parser,
which this dist no longer uses, Pod::PlainText and its dependencies were
copied directly into this dist for the test to use.

Convert the test to use Pod::Text instead, which provides very similar
output. This allows removing the Pod::Parser modules, as well as
simplifying the test greatly.